[Silver V] Title: 타일 장식물, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.cc
@@ -2,8 +2,8 @@
 
 using namespace std;
 
-int dp_perimeter[81];
-int dp_line[81];
+long long dp_perimeter[81];
+long long dp_line[81];
 
 int main() {
     int n;
